requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.13
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-grappelli==4.0.1
django-timezone-field==7.0
djangorestframework==3.15.2
idna==3.7
kombu==5.4.0
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.22.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
telebot==0.0.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13

django-appconf==1.0.4
